move zerocopy tests into tests file
