+++ src/tests.rs
@@ -254,3 +254,29 @@ async fn unlink_closed() {
 
     object.recv(move |s| assert_eq!(s, expected)).await.unwrap();
 }
+
+impl OwningMessage for String {
+    type NonOwning<'a> = &'a str;
+
+    fn to_non_owned(&self) -> Self::NonOwning<'_> {
+        self.as_str()
+    }
+}
+
+impl<'a> NonOwningMessage<'a> for &'a str {
+    type Owning = String;
+
+    fn to_owned(self) -> Self::Owning {
+        self.to_string()
+    }
+}
+
+#[tokio::test]
+async fn alias_sent_on_pending_recv() {
+    let (tx, rx) = channel::<String>();
+    let sent = "shared string";
+    let test = |received: &str| sent.as_ptr() == received.as_ptr();
+    let join_rx = tokio::spawn(async move { rx.recv(test).await });
+    tx.send(sent).unwrap().await;
+    assert!(join_rx.await.unwrap().unwrap());
+}
